+++ src/main/java/it/gov/pagopa/pu/organization/service/TaxonomyService.java
@@ -1,20 +1,10 @@
 package it.gov.pagopa.pu.organization.service;
 
 import it.gov.pagopa.pu.organization.model.DistinctOrganizationTypeDTO;
-import it.gov.pagopa.pu.organization.repository.TaxonomyRepository;
 import java.util.List;
 import org.springframework.data.domain.Sort;
-import org.springframework.stereotype.Service;
 
-@Service
-public class TaxonomyService {
-  private TaxonomyRepository taxonomyRepository;
-  public TaxonomyService(TaxonomyRepository taxonomyRepository){
-    this.taxonomyRepository = taxonomyRepository;
-  }
-
-  public List<DistinctOrganizationTypeDTO> getDistinctOrganizationType(Sort sort){
-    return taxonomyRepository.findDistinctOrganizationTypes(sort);
-  }
+public interface TaxonomyService {
+  List<DistinctOrganizationTypeDTO> getDistinctOrganizationType(Sort sort);
 
 }
